cli/ruby-debug/interface.rb: add a verbose mode to show what commands are getting run.
runner.sh: add the ability to set the rdebug script to run  based on environment
variable RDEBUG

Both of the above are in support of adding the first integration test
by running rdebug and comparing output produced. The first test is of
commands in stepping.rb (step, next, finish). It is still a little
hoaky and needs more work, but another step in the right direction.



git-svn-id: http://ruby-debug.rubyforge.org/svn/trunk@305 453b2852-3d18-0410-866e-d09c099698e4