From Anders Lindgren: * `gud-comint-buffer' is now buffer local in all secondary buffers, hopefully this will make `rdebug' more independent of other processes like `gdb'.
* The restore window configuration variable now has three possible
values, when set to :many (which is the new default) the window
configuration is only restored when `rdebug-many-windows' is active.

* The `edit variable' command, when reading from the minibuffer, uses
the old value as the initial content. Mouse-1 is no longer bound to
`edit' since that prevents a simple copy, instead mouse-2 and mouse-3
is used.

* More robust window configuration restore system. Trace support for
window configuration state changes added.


git-svn-id: http://ruby-debug.rubyforge.org/svn/trunk@391 453b2852-3d18-0410-866e-d09c099698e4