Moved handling of command output out from gud-rdebug-marker-filter. It is now handled by rdebug-cmd-XXX (which used to be located in rdebug-info.el.) Tooltip handling is now more robust, e.g. it works when more than one annotation is received at once, or if the annotation arrives in split packages.
In case the external process echoes the command, it is stipped away
before it is presented as a tooltip.

Test of filter function reworked. New cases added to handle command
output. All filter tests are now tested twice, once when the entire
block is sent at once, one where the block is sent in character by
character.


git-svn-id: http://ruby-debug.rubyforge.org/svn/trunk@692 453b2852-3d18-0410-866e-d09c099698e4